Refactored for separation of concerns
